extract init_db stuff.
